rework cast to avoid cast_possible_truncation lint
in CreateUploadData::new
warning: casting `u64` to `usize` may truncate the value on targets with 32-bit wide pointers..
The original value is u32, so :
Convert u32 data to usize to use in indexing and usize to u64 for cursor position.
Instead of convert u32 data to u64 and tried after to convert the u64 in usize.

+++ src/model/cargo.rs
@@ -317,14 +317,14 @@ impl CrateUploadData {
     pub fn new(buffer: &[u8]) -> Result<Self, ApiError> {
         let mut cursor = Cursor::new(buffer);
         // read the metadata
-        let metadata_length = u64::from(cursor.read_u32::<LittleEndian>()?);
-        let metadata_buffer = &buffer[4..((4 + metadata_length) as usize)];
+        let metadata_length = cursor.read_u32::<LittleEndian>()? as usize;
+        let metadata_buffer = &buffer[4..(4 + metadata_length)];
         let metadata = serde_json::from_slice(metadata_buffer)?;
         // read the content
-        cursor.set_position(4 + metadata_length);
+        cursor.set_position(4 + metadata_length as u64);
         let content_length = cursor.read_u32::<LittleEndian>()? as usize;
         let mut content = vec![0_u8; content_length];
-        content.copy_from_slice(&buffer[((4 + metadata_length + 4) as usize)..]);
+        content.copy_from_slice(&buffer[(4 + metadata_length + 4)..]);
         Ok(Self { metadata, content })
     }
